Duplicate cache path determination

+++ pathcache/pathcache.py
@@ -17,43 +17,45 @@ class PathCacheManager(object):
     def pathstr(self, path: str, foldercontext=False, refresh=False) -> str:
         # get the absolute path
         abs_path = os.path.abspath(path)
-
+        
         result_path = abs_path
+        # if path not in cache path
+        if abs_path[:len(self.cache_root_folder_path)]!=self.cache_root_folder_path:
+                
+            if abs_path not in self.cache_dict:
+                if abs_path[0] == '/':
+                    self.cache_dict[abs_path] = os.path.join(
+                        self.cache_root_folder_path, abs_path[1:])
+                elif abs_path.split(':')[1][0] == '/' or abs_path.split(':')[1][0] == '\\':
+                    self.cache_dict[abs_path] = os.path.join(
+                        self.cache_root_folder_path, abs_path.split(':')[0], abs_path.split(':')[1][1:])
+                refresh = True
+            result_path = self.cache_dict[abs_path]
 
-        if abs_path not in self.cache_dict:
-            if abs_path[0] == '/':
-                self.cache_dict[abs_path] = os.path.join(
-                    self.cache_root_folder_path, abs_path[1:])
-            elif abs_path.split(':')[1][0] == '/' or abs_path.split(':')[1][0] == '\\':
-                self.cache_dict[abs_path] = os.path.join(
-                    self.cache_root_folder_path, abs_path.split(':')[0], abs_path.split(':')[1][1:])
-            refresh = True
-        result_path = self.cache_dict[abs_path]
-
-        if refresh:
-            if os.path.isfile(abs_path):
-                if foldercontext:
-                    if os.path.exists(os.path.split(self.cache_dict[abs_path])[0]):
-                        os.rmdir(os.path.split(self.cache_dict[abs_path])[0])
-                    shutil.copytree(os.path.split(abs_path)[
-                        0], os.path.split(self.cache_dict[abs_path])[0])
-                else:
-                    os.makedirs(os.path.split(self.cache_dict[abs_path])[0], exist_ok=True)
-
-                if os.path.exists(abs_path):
-                    if os.path.exists(self.cache_dict[abs_path]):
-                        if not filecmp.cmp(abs_path, self.cache_dict[abs_path]):
-                            shutil.copy(abs_path, self.cache_dict[abs_path])
+            if refresh:
+                if os.path.isfile(abs_path):
+                    if foldercontext:
+                        if os.path.exists(os.path.split(self.cache_dict[abs_path])[0]):
+                            os.rmdir(os.path.split(self.cache_dict[abs_path])[0])
+                        shutil.copytree(os.path.split(abs_path)[
+                            0], os.path.split(self.cache_dict[abs_path])[0])
                     else:
-                        shutil.copy(abs_path, self.cache_dict[abs_path])
-            else:
-                if foldercontext:
-                    if os.path.exists(self.cache_dict[abs_path]):
-                        shutil.rmtree(self.cache_dict[abs_path])
-                    shutil.copytree(
-                        abs_path, self.cache_dict[abs_path])
+                        os.makedirs(os.path.split(self.cache_dict[abs_path])[0], exist_ok=True)
+
+                    if os.path.exists(abs_path):
+                        if os.path.exists(self.cache_dict[abs_path]):
+                            if not filecmp.cmp(abs_path, self.cache_dict[abs_path]):
+                                shutil.copy(abs_path, self.cache_dict[abs_path])
+                        else:
+                            shutil.copy(abs_path, self.cache_dict[abs_path])
                 else:
-                    os.makedirs(self.cache_dict[abs_path], exist_ok=True)
+                    if foldercontext:
+                        if os.path.exists(self.cache_dict[abs_path]):
+                            shutil.rmtree(self.cache_dict[abs_path])
+                        shutil.copytree(
+                            abs_path, self.cache_dict[abs_path])
+                    else:
+                        os.makedirs(self.cache_dict[abs_path], exist_ok=True)
 
         return result_path
